Play video instead of scrub-syncing to audio
Scrubbing is very jank, probably something to do with hardware
acceleration only working during normal playback.

Right now the audio buffers are prebuffered 100%, so in theory,
nothing should get out of sync. But in the future, I think the
audio have to be streamed, so that will require some sync work.